% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgpd.R
\name{dgpd}
\alias{dgpd}
\title{Probability distribution function of GPD}
\usage{
dgpd(x, mu = 0, sigma = 1, xi = 1, log = F)
}
\arguments{
\item{x}{numeric predictor matrix.}

\item{y}{numeric response vector.}

\item{a}{shape parameter for gamma prior distributions.}

\item{b}{rate parameter for gamma prior distributions.}

\item{mu.start}{initial value for mu parameter.}

\item{beta.prior.mean}{mu parameter for normal prior distributions.}

\item{beta.prior.sd}{standard deviation for normal prior distributions.}

\item{iters}{number of iterations for the Markov chain to run.}

\item{burn}{numeric burn-in length.}

\item{nthin}{numeric thinning rate.}

\item{plots}{logical operator. \code{TRUE} to output plots.}

\item{progress.bar}{logical operator. \code{TRUE} to print progress bar.}
}
\value{

}
\description{
\code{zero_nb} is used to fit zero-inflated 
negative binomial regression models to count data via Bayesian inference.
}
\details{

}
\examples{

}
\author{
Dr. Earvin Balderama <\email{ebalderama@luc.edu}>
}

