% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_zinb.R
\name{loglik_zinb}
\alias{loglik_zinb}
\title{Zero-inflated Negative Binomial Data Likelihood}
\usage{
loglik_zinb(y, z, mu, size, p)
}
\arguments{
\item{y}{numeric response vector.}

\item{z}{vector of binary operators. \code{z == 0} for observations 
considered belnging to the negative binomial distribution, \code{z == 1} 
for observations considered to be 'extra' zeros.}

\item{mu}{current value for the negative binomial likelihood mu parameter.}

\item{size}{size parameter for negative binomial distribution.}

\item{p}{vector of 'extra' zero-count probabilities.}
}
\value{
The log-likelihood of the zero-inflated negative binomial fit 
for the current iteration of the MCMC algorithm.
}
\description{
Data likelihood fuction for zero-inflated negative binomial 
model regression using \code{\link{zero_nb}}.
}
\details{

}
\examples{

}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Earvin Balderama <\email{ebalderama@luc.edu}>
}
\seealso{
\code{\link{zero_nb}}
}

