% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_zip.R
\name{loglik_zip}
\alias{loglik_zip}
\title{Zero-inflated Poisson Data Likelihood}
\usage{
loglik_zip(y, z, lam, p)
}
\arguments{
\item{y}{numeric response vector.}

\item{z}{vector of binary operators. \code{z == 0} for observations 
considered belnging to the negative binomial distribution, \code{z == 1} 
for observations considered to be 'extra' zeros.}

\item{lam}{current value for the Poisson likelihood lambda parameter.}

\item{p}{vector of 'extra' zero-count probabilities.}
}
\value{
The log-likelihood of the zero-inflated Poisson fit for the
current iteration of the MCMC algorithm.
}
\description{
Data likelihood fuction for zero-inflated Poisson model 
regression using \code{\link{zero_poisson}}.
}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Earvin Balderama <\email{ebalderama@luc.edu}>
}
\seealso{
\code{\link{zero_poisson}}
}

