% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpd.R
\name{mgpd}
\alias{mgpd}
\title{Probability distribution function of discrete GPD}
\usage{
mgpd(x, mu = 0, sigma = 1, xi = 1, log = F)
}
\arguments{
\item{x}{numeric predictor matrix.}

\item{y}{numeric response vector.}

\item{hurd}{numeric threshold for 'extreme' observations of two-hurdle models. 
\code{Inf} for one-hurdle models.}

\item{dist}{character specification of response distribution.}

\item{dist.2}{character specification of response distribution for 
'extreme' observations of two-hurdle models.}

\item{control}{list of parameters for controlling the fitting process, 
specified by \code{\link{hurdle_control}}.}

\item{iters}{number of iterations for the Markov chain to run.}

\item{burn}{numeric burn-in length.}

\item{nthin}{numeric thinning rate.}

\item{plots}{logical operator. \code{TRUE} to output plots.}

\item{progress.bar}{logical operator. \code{TRUE} to print progress bar.}
}
\value{

}
\description{
\code{hurdle} is used to fit single or 
double-hurdle regression models to count data via Bayesian inference.
}
\details{

}
\examples{

}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Dr. Earvin Balderama <\email{ebalderama@luc.edu}>
}

