% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE.R
\name{PE}
\alias{PE}
\title{Extreme Count Probability Likelihood}
\usage{
PE(p, q, log = T)
}
\arguments{
\item{p}{vector of zero-count probabilities.}

\item{q}{vector of 'extreme' count probabilities.}

\item{log}{logical operator. If \code{TRUE}, probabilities \code{p} 
and \code{q} are given as \code{log(p)}, \code{log(q)}.}
}
\value{
A vector of probabilities.
}
\description{
\code{PE} is used to calculate the likelihood of a user-defined
'extreme' value count observation in a double-hurdle regression model.
}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Earvin Balderama <\email{ebalderama@luc.edu}>
}
\seealso{
\code{\link{hurdle}}
}

