% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpd.R
\name{pgpd}
\alias{pgpd}
\title{Cumulative distribution function of GPD}
\usage{
pgpd(q, mu = 0, sigma = 1, xi = 1, lower.tail = T)
}
\arguments{
\item{q}{vector of 'extreme' count probabilities.}

\item{p}{vector of zero-count probabilities.}

\item{log}{logical operator. If \code{TRUE}, probabilities \code{p} 
and \code{q} are given as \code{log(p)}, \code{log(q)}.}
}
\value{

}
\description{
\code{PE} is used to calculate the likelihood of a user-defined
'extreme' value count observation in a double-hurdle regression model.
}
\details{

}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Dr. Earvin Balderama <\email{ebalderama@luc.edu}>
}
\seealso{
\code{\link{hurdle}}
}

