% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT.R
\name{PT}
\alias{PT}
\title{Typical Count Probability Likelihood}
\usage{
PT(p, q, log = T)
}
\arguments{
\item{p}{vector of zero-count probabilities.}

\item{q}{vector of 'typical' count probabilities.}

\item{log}{logical operator. If \code{TRUE}, probabilities \code{p} 
and \code{q} are given as \code{log(p)}, \code{log(q)}.}
}
\value{
A vector of probabilities.
}
\description{
\code{PT} is used to calculate the likelihood of a user-defined
'typical' value count observation in a double-hurdle regression model.
}
\details{

}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Earvin Balderama <\email{ebalderama@luc.edu}>
}
\seealso{
\code{\link{hurdle}}
}

