% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_zinb.R
\name{loglik_zinb}
\alias{loglik_zinb}
\title{Define zero-inflated negative binomial data loglikelihood at z = 1, z = 0}
\usage{
loglik_zinb(y, z, mu, size, p)
}
\arguments{
\item{y}{numeric response vector.}

\item{x}{numeric predictor matrix.}

\item{hurd}{numeric threshold for 'extreme' observations of two-hurdle models. 
\code{Inf} for one-hurdle models.}

\item{dist}{character specification of response distribution.}

\item{dist.2}{character specification of response distribution for 
'extreme' observations of two-hurdle models.}

\item{control}{list of parameters for controlling the fitting process, 
specified by \code{\link{hurdle_control}}.}

\item{iters}{number of iterations for the Markov chain to run.}

\item{burn}{numeric burn-in length.}

\item{nthin}{numeric thinning rate.}

\item{plots}{logical operator. \code{TRUE} to output plots.}

\item{progress.bar}{logical operator. \code{TRUE} to print progress bar.}
}
\value{

}
\description{
\code{hurdle} is used to fit single or 
double-hurdle regression models to count data via Bayesian inference.
}
\details{

}
\examples{

}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Dr. Earvin Balderama <\email{ebalderama@luc.edu}>
}

