% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PZ.R
\name{PZ}
\alias{PZ}
\title{Zero Count Probability Likelihood}
\usage{
PZ(p, log = T)
}
\arguments{
\item{p}{vector of zero-count probabilities.}

\item{log}{logical operator. If \code{TRUE}, probabilities \code{p} 
and \code{q} are given as \code{log(p)}, \code{log(q)}.}
}
\value{
A vector of probabilities.
}
\description{
\code{PZ} is used to calculate the likelihood of a 
zero-value count observation in a single or double-hurdle regression model.
}
\author{
Taylor Trippe <\email{ttrippe@luc.edu}> \cr
Earvin Balderama <\email{ebalderama@luc.edu}>
}
\seealso{
\code{\link{hurdle}}
}

